+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- Fixes a bug preventing local extension instances from being updated to uploaded versions.
+- Added more helpful error messages for the Firebase Hosting GitHub Action (#5749)
+- Upgrade Firestore emulator to 1.17.4
+- Web Frameworks will no longer try to deploy unsupported versions of NodeJS to Cloud Functions (#5733)